player(перенесла в untils функцию)
